gh-1453 Added ledger hash
