+++ Multisig/UI/Settings/OwnerKeyManagement/LedgerOwnerKey/LedgerPendingConfirmationViewController.swift
@@ -12,21 +12,20 @@ class LedgerPendingConfirmationViewController: UIViewController {
     @IBOutlet private weak var bottomView: UIView!
     @IBOutlet private weak var headerLabel: UILabel!
     @IBOutlet private weak var descriptionLabel: UILabel!
-    // NOTE: hidden until further releases
     @IBOutlet private weak var safeTxHashLabel: UILabel!
     @IBOutlet private weak var cancelButton: UIButton!
 
     var headerText = "Confirm Transaction"
-    var safeTxHash: HashString!
+    var ledgerHash: String!
     var onClose: (() -> Void)?
 
     @IBAction private func cancel(_ sender: Any) {
         close()
     }
 
-    convenience init(safeTxHash: HashString, headerText: String? = nil) {
+    convenience init(ledgerHash: String, headerText: String? = nil) {
         self.init(nibName: nil, bundle: nil)
-        self.safeTxHash = safeTxHash
+        self.ledgerHash = ledgerHash
         if let headerText = headerText {
             self.headerText = headerText
         }
@@ -47,7 +46,7 @@ class LedgerPendingConfirmationViewController: UIViewController {
         headerLabel.text = headerText
         headerLabel.setStyle(.headline)
         descriptionLabel.setStyle(.callout)
-        safeTxHashLabel.attributedText = safeTxHash.highlighted
+        safeTxHashLabel.attributedText = highlightedLedgerHash
         cancelButton.setText("Cancel", .plain)
 
         modalTransitionStyle = .crossDissolve
@@ -56,5 +55,24 @@ class LedgerPendingConfirmationViewController: UIViewController {
     private func close() {
         dismiss(animated: true)
     }
-}
 
+    var highlightedLedgerHash: NSAttributedString {
+        let value = ledgerHash!
+        let style = GNOTextStyle.tertiary
+        let attributedString = NSMutableAttributedString(
+            string: value,
+            attributes: [
+                .font: UIFont.gnoFont(forTextStyle: style),
+                .foregroundColor: style.color!,
+                .kern: -0.41
+            ]
+        )
+        // first 4 digits
+        attributedString.addAttribute(
+            .foregroundColor, value: UIColor.primaryLabel, range: NSRange(location: 0, length: 4))
+        // last 4 digits
+        attributedString.addAttribute(
+            .foregroundColor, value: UIColor.primaryLabel, range: NSRange(location: value.count - 4, length: 4))
+        return attributedString
+    }
+}
