+++ Multisig/UI/Settings/OwnerKeyManagement/LedgerOwnerKey/LedgerPendingConfirmationViewController.xib
@@ -25,10 +25,10 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="AyD-YA-oAT">
-                    <rect key="frame" x="0.0" y="496.5" width="414" height="399.5"/>
+                    <rect key="frame" x="0.0" y="445.5" width="414" height="450.5"/>
                     <subviews>
                         <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="center" spacing="22" translatesAutoresizingMaskIntoConstraints="NO" id="TtX-3G-9mc">
-                            <rect key="frame" x="30" y="24" width="354" height="325.5"/>
+                            <rect key="frame" x="30" y="24" width="354" height="376.5"/>
                             <subviews>
                                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Confirm Transaction" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="4Vx-j0-agE">
                                     <rect key="frame" x="100" y="0.0" width="154.5" height="20.5"/>
@@ -49,14 +49,14 @@
                                         <constraint firstAttribute="height" constant="122" id="hCV-xM-a4Q"/>
                                     </constraints>
                                 </imageView>
-                                <label hidden="YES" opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="0x9342446f11d3c46bc1f838100fef789c174a7db4147635d7f5709fc2ff72b2ba" textAlignment="center" lineBreakMode="middleTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" minimumScaleFactor="0.69999999999999996" translatesAutoresizingMaskIntoConstraints="NO" id="HiR-2U-6CV">
-                                    <rect key="frame" x="-124" y="258.5" width="602" height="0.0"/>
+                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="0x9342446f11d3c46bc1f838100fef789c174a7db4147635d7f5709fc2ff72b2ba" textAlignment="center" lineBreakMode="middleTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" minimumScaleFactor="0.69999999999999996" translatesAutoresizingMaskIntoConstraints="NO" id="HiR-2U-6CV">
+                                    <rect key="frame" x="0.0" y="269.5" width="354" height="29"/>
                                     <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                     <nil key="textColor"/>
                                     <nil key="highlightedColor"/>
                                 </label>
                                 <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="HP4-d5-dLu">
-                                    <rect key="frame" x="16" y="269.5" width="322" height="56"/>
+                                    <rect key="frame" x="16" y="320.5" width="322" height="56"/>
                                     <constraints>
                                         <constraint firstAttribute="height" constant="56" id="wMS-Ae-e1h"/>
                                     </constraints>
