+++ Multisig/UI/WalletConnect/WCTransactionConfirmationViewController.swift
@@ -419,7 +419,7 @@ extension WCTransactionConfirmationViewController: SelectLedgerDeviceDelegate {
               let metadata = keyInfo.metadata,
               let ledgerKeyMetadata = KeyInfo.LedgerKeyMetadata.from(data: metadata) else { return }
 
-        let pendingConfirmationVC = LedgerPendingConfirmationViewController(safeTxHash: transaction.safeTxHash!)
+        let pendingConfirmationVC = LedgerPendingConfirmationViewController(ledgerHash: transaction.hardwareWalletHash)
         pendingConfirmationVC.modalPresentationStyle = .popover
 
         // dismiss Select Ledger Device screen and presend Ledger Pending Confirmation overlay
